Fixed Bug #3166
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@4525 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/hruler.h
@@ -52,7 +52,7 @@ public:
 	void mouseMoveEvent(QMouseEvent *m);
 	void UpdateTabList();
 	void setItemPosition(double pos, double width);
-	int offs;
+	double offs;
 	int Markp;
 	int oldMark;
 	bool Mpressed;
